--- src/GEM_README.md
@@ -1,30 +1,0 @@
-# Truthy Gem
-
-### What is truthy anyway?
-
-Well, think of a truth table you just invented, you put your 1s and 0s and **Truthy** figures out which formula works
-for the table.
-
-| A   | B   | C   |
-|-----|-----|-----|
-| 0   | 0   | 1   |
-| 1   | 0   | 1   |
-| 1   | 1   | 0   |
-| 0   | 1   | 0   |
-
-If you create the above table with **Truthy**, it will know what to do so that your table makes sense.
-Also, **Truthy** has implementations for the following logical gates: `and`, `or`, `not`,
-`xor`, `nand`, `nor`, and `xnor`.
-
-### Where can I use it?
-
-- Access matrices are a good example;
-- Other examples are just day-to-day comparisons you do with booleans in your code.
-
-Let your imagination make good use of Truthy.
-
-### Where do I learn how to use it?
-
-Check the documentation [here](https://github.com/roberwil/truthy_gem/blob/main/README.md).
-Remember, **Truthy** is the way.
-
