+++ src/Rakefile
@@ -1,7 +1,10 @@
+require "bundler/gem_tasks"
 require 'rake/testtask'
 
 Rake::TestTask.new do |t|
   t.libs << "test"
+  t.libs << "lib"
+  t.test_files = FileList["test/**/*_test.rb"]
   #t.pattern = "test/*_test.rb"
 end
