+++ src/ada_truthy.gemspec
@@ -1,6 +1,8 @@
+require_relative 'lib/ada_truthy/version'
+
 Gem::Specification.new do |s|
   s.name                  = "ada_truthy"
-  s.version               = "1.0.0"
+  s.version               = AdaTruthy::VERSION
   s.summary               = "Create truth tables and operate with logical gates."
   s.description           = "Create truth tables and let truthy handle the annoying part. Also, operate with logical gates."
   s.authors               = "Roberto W. P. Ribeiro"
@@ -15,5 +17,4 @@ Gem::Specification.new do |s|
   s.homepage              = "https://github.com/roberwil/truthy_gem"
   s.license               = "MIT"
   s.required_ruby_version = '>= 2.7.7'
-  s.extra_rdoc_files      = ['GEM_README.md']
 end
