+++ src/lib/ada_truthy/truth_table/truthy_exception.rb
@@ -1,0 +1,4 @@
+module AdaTruthy
+  class TruthyException < StandardError
+  end
+end
