+++ src/lib/ada_truthy/version.rb
@@ -1,0 +1,3 @@
+module AdaTruthy
+  VERSION = "1.0.0"
+end
