--- src/lib/truth_table/truthy_exception.rb
@@ -1,2 +1,0 @@
-class TruthyException < StandardError
-end
