+++ src/test/test_helper.rb
@@ -1,0 +1,3 @@
+$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
+require 'ada_truthy'
+require "minitest/autorun"
